THOR-1430 Moved createRelationFilterList from FilterService into AbstractFilter.
