+++ src/app/library/core/models/filters.ts
@@ -444,6 +444,55 @@ export abstract class AbstractFilter {
     ): AbstractFilter;
 
     /**
+     * Creates and returns the relation filter list for the filter.
+     *
+     * @arg {Dataset} dataset
+     * @return {AbstractFilter[]}
+     */
+    public createRelationFilterList(dataset: Dataset): AbstractFilter[] {
+        let filterDataSourceList: FilterDataSource[] = FilterUtil.createFilterDataSourceListFromConfig(this.toConfig(), true);
+
+        return dataset.getRelations().reduce((returnList, relation) => {
+            let relationFilterList: AbstractFilter[] = [];
+
+            // Assume that each item within the relation list is a nested list with the same length.
+            // EX:  [[x1, y1], [x2, y2], [x3, y3]]
+            if (relation.length && relation[0].length === filterDataSourceList.length) {
+                let equivalentRelationList: FieldKey[][] = relation.filter((relationFieldKeyList) =>
+                    // Each item within the relationFieldKeyList must be equivalent to a FilterDataSource.
+                    relationFieldKeyList.every((relationFieldKey) => filterDataSourceList.some((filterDataSource) =>
+                        this._isRelationEquivalent(relationFieldKey, filterDataSource))) &&
+                    // Each FilterDataSource must be equivalent to an item within the relationFieldKeyList.
+                    filterDataSourceList.every((filterDataSource) => relationFieldKeyList.some((relationFieldKey) =>
+                        this._isRelationEquivalent(relationFieldKey, filterDataSource))));
+
+                // The length of equivalentRelationList should be either 0 or 1.
+                if (equivalentRelationList.length) {
+                    // Create new relation filters.
+                    relation.forEach((relationFieldKeyList) => {
+                        // Do not create a relation that is the same as the original filter.
+                        if (relationFieldKeyList !== equivalentRelationList[0]) {
+                            let relationFilter: AbstractFilter = this.createRelationFilter(equivalentRelationList[0],
+                                relationFieldKeyList, dataset);
+                            relationFilterList.push(relationFilter);
+                        }
+                    });
+
+                    // Save sibling relation filter IDs in the new relation filters.
+                    relationFilterList.concat(this).forEach((outerFilter) => {
+                        relationFilterList.concat(this).forEach((innerFilter) => {
+                            if (outerFilter.id !== innerFilter.id) {
+                                outerFilter.relations.push(innerFilter.id);
+                            }
+                        });
+                    });
+                }
+            }
+            return returnList.concat(relationFilterList);
+        }, [] as AbstractFilter[]);
+    }
+
+    /**
      * Returns if this filter affects a search in the given datastore/database/table.
      *
      * @arg {string} datastore
@@ -500,6 +549,19 @@ export abstract class AbstractFilter {
      * @abstract
      */
     public abstract isEquivalentToFilter(filter: AbstractFilter): boolean;
+
+    /**
+     * Returns if the given field is equivalent to the given data source.
+     *
+     * @arg {FieldKey} fieldKey
+     * @arg {FilterDataSource} filterDataSource
+     * @return {boolean}
+     * @private
+     */
+    private _isRelationEquivalent(fieldKey: FieldKey, filterDataSource: FilterDataSource): boolean {
+        return !!(fieldKey.datastore === filterDataSource.datastore && fieldKey.database === filterDataSource.database &&
+            fieldKey.table === filterDataSource.table && fieldKey.field === filterDataSource.field);
+    }
 
     /**
      * Returns the filtered values for the filter object.
